Scroll callouts using the same logic as tours.
The biggest change is making `adjustWindowScroll` visible to the
callout manager, where as previously adjustWindowScroll was a private
method internal to the Hopscotch class.

There are probably more elegant ways of doing this, e.g. extracting
it into utils instead, but this was most straight for me.

Also, `adjustWindowScroll` now takes paramter, `elem` (bad name),
to pass to getStepTarget, instead of the previous behavior of
calling getCurrStep(), which assumed a tour was in progress.

Fixes #119.